Add support for int16/int32 to CRSF parameters

+++ src/lib/CrsfProtocol/CRSFParameters.h
@@ -132,7 +132,7 @@ struct elrsStatusParameter
             LUA_FIELD_HIDE(fld)      \
     }
 
-typedef std::function<void(propertiesCommon *item, uint8_t arg)> parameterHandlerCallback;
+typedef std::function<void(propertiesCommon *item, int32_t arg)> parameterHandlerCallback;
 
 uint8_t findSelectionLabel(const selectionParameter *parameter, char *outArray, uint8_t value);
